+++ CHANGES.md
@@ -83,6 +83,7 @@ Change Log
 * Provide a fallback name for an `ArcGisServerCatalogItem`
 * Ensure `CesiumTileLayer.getTileUrl` returns a string.
 * Adds methods `removeModelReferences` to Terria & ViewState for unregistering and removing models from different parts of the UI.
+* Add basic support for various error provider services, implementing support for Rollbar. 
 
 ### Next Release
 * Fix draggable workbench/story items with translation HOC